# apps/api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/admin/concert":{"post":{"operationId":"AdminConcertController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateConcertDto"}}}},"responses":{"201":{"description":"Concert created successfully"},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"summary":"Create a new concert","tags":["AdminConcert"]},"get":{"operationId":"AdminConcertController_findAll","parameters":[],"responses":{"200":{"description":"Concerts retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetAdminConcertDto"}}}}},"404":{"description":"Concerts not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"summary":"Get all concerts for admin","tags":["AdminConcert"]}},"/admin/concert/{id}":{"delete":{"operationId":"AdminConcertController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Concert deleted successfully"},"404":{"description":"Concert not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"summary":"Delete a concert by ID","tags":["AdminConcert"]}},"/admin/concert/reservation/history":{"get":{"operationId":"AdminConcertController_findAllHistory","parameters":[],"responses":{"200":{"description":"Reservation history retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetReservationHistoryDto"}}}}},"404":{"description":"Reservation history not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"summary":"Get all reservation history","tags":["AdminConcert"]}},"/admin/concert/dashboard":{"get":{"operationId":"AdminConcertController_getDashboard","parameters":[],"responses":{"200":{"description":"Dashboard information retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAdminDashboardDto"}}}}},"summary":"Get admin dashboard information","tags":["AdminConcert"]}},"/user/concert":{"get":{"operationId":"UserConcertController_findAll","parameters":[],"responses":{"200":{"description":"Concerts retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetUserConcertDto"}}}}},"404":{"description":"Concerts not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"summary":"Get all concerts for user","tags":["UserConcert"]}},"/user/concert/{id}/reservation/reserve":{"post":{"operationId":"UserConcertController_reserve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Seat reserved successfully"},"400":{"description":"No seats available","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Concert not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"summary":"Reserve a seat for a concert","tags":["UserConcert"]}},"/user/concert/{id}/reservation/cancel":{"post":{"operationId":"UserConcertController_cancelReservation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Reservation canceled successfully"},"400":{"description":"Reservation is already cancelled","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Concert or reservation not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"summary":"Cancel a reservation for a concert","tags":["UserConcert"]}}},"info":{"title":"Concert API","description":"The Concert API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateConcertDto":{"type":"object","properties":{"name":{"type":"string"},"totalSeat":{"type":"number"},"description":{"type":"string"}},"required":["name","totalSeat","description"]},"ErrorResponse":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"oneOf":[{"type":"string"},{"type":"array","items":{"type":"string"}}]},"error":{"type":"string","example":"Bad Request"}},"required":["statusCode","message","error"]},"GetAdminConcertDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"string"},"totalSeat":{"type":"number"},"currentTotalSeat":{"type":"number"}},"required":["id","name","description","totalSeat","currentTotalSeat"]},"GetReservationHistoryDto":{"type":"object","properties":{"id":{"type":"number"},"concertName":{"type":"string"},"userId":{"type":"number"},"status":{"type":"string","enum":["RESERVE","CANCEL","NONE"]},"createdAt":{"format":"date-time","type":"string"}},"required":["id","concertName","userId","status","createdAt"]},"GetAdminDashboardDto":{"type":"object","properties":{"totalSeats":{"type":"number"},"totalReserveReservation":{"type":"number"},"totalCancelReservation":{"type":"number"}},"required":["totalSeats","totalReserveReservation","totalCancelReservation"]},"GetUserConcertDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"string"},"totalSeat":{"type":"number"},"currentTotalSeat":{"type":"number"},"reservationStatus":{"type":"string","enum":["RESERVE","CANCEL","NONE"]}},"required":["id","name","description","totalSeat","currentTotalSeat","reservationStatus"]}}}}